添加消息提醒

+++ src/main/java/com/buguagaoshu/community/model/Notification.java
@@ -41,5 +41,5 @@ public class Notification {
     /**
      * 通知状态
      */
-    private long status;
+    private int status;
 }
